backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anthropic==0.87.0",
    "anyio==4.13.0",
    "beautifulsoup4==4.14.3",
    "blinker==1.9.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "colorama==0.4.6",
    "comtypes==1.4.16",
    "cryptography==47.0.0",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "exceptiongroup==1.3.1",
    "fastapi==0.136.1",
    "flask==3.1.3",
    "google-ai-generativelanguage",
    "google-api-core",
    "google-api-python-client==2.194.0",
    "google-auth==2.49.2",
    "google-auth-httplib2==0.3.1",
    "google-genai==1.73.1",
    "google-generativeai==0.8.6",
    "googleapis-common-protos==1.74.0",
    "greenlet==3.4.0",
    "groq==0.37.1",
    "grpcio==1.80.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.2",
    "httpx==0.28.1",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "keyboard==0.13.5",
    "langchain==1.2.14",
    "langchain-anthropic==1.4.0",
    "langchain-core==1.2.23",
    "langchain-groq==1.1.2",
    "langgraph==1.1.4",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.12",
    "langsmith==0.7.23",
    "lxml==6.1.0",
    "markupsafe==3.0.3",
    "mouseinfo==0.1.3",
    "nltk==3.9.4",
    "numpy==2.2.6",
    "opencv-python==4.13.0.92",
    "orjson==3.11.8",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pillow==12.2.0",
    "proto-plus==1.27.2",
    "protobuf==5.29.6",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pyautogui==0.9.54",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygetwindow==0.0.9",
    "pymsgbox==2.0.1",
    "pyparsing==3.3.2",
    "pyperclip==1.11.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "python-docx==1.2.0",
    "python-dotenv==1.2.2",
    "pyttsx3==2.99",
    "pytweening==1.2.0",
    "pywhatkit==5.4",
    "pyyaml==6.0.3",
    "regex==2026.3.32",
    "requests==2.33.1",
    "requests-toolbelt==1.0.0",
    "sarvam==0.0.0",
    "sarvamai==0.1.27",
    "sniffio==1.3.1",
    "sounddevice==0.5.5",
    "soundfile==0.13.1",
    "soupsieve==2.8.3",
    "speechrecognition==3.15.2",
    "sqlalchemy==2.0.49",
    "starlette==1.0.0",
    "tenacity==9.1.4",
    "textblob==0.20.0",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uvicorn==0.46.0",
    "websockets==16.0",
    "werkzeug==3.1.8",
    "wikipedia==1.4.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
